Abstracted Fret to its own component
