+++ src/components/Guitar/subcomponents/Fret/Fret.scss
@@ -1,0 +1,58 @@
+.fret {
+  --background-color: saddlebrown;
+  all: unset;
+  display: grid;
+  place-content: center;
+  flex: 1 1 0px;
+  height: 100%;
+  border: white solid 1px;
+  color: var(--text-dark-low-emphasis);
+  background-color: var(--background-color);
+  cursor: pointer;
+
+  &:hover {
+    background-color: hsl(from var(--background-color) h s 50);
+  }
+
+  &:first-child {
+    --background-color: black;
+    color: var(--text-light-low-emphasis-locked);
+  }
+
+  :global([data-theme="dark"]) & {
+    --background-color: black;
+
+    &:first-child {
+      --background-color: rgb(150, 150, 150);
+      color: var(--text-dark-low-emphasis-locked);
+
+      &:hover {
+        background-color: hsl(from var(--background-color) h s 80);
+      }
+    }
+  }
+
+  &__indicator {
+    display: grid;
+    place-content: center;
+    width: 1.5rem;
+    height: 1.5rem;
+    border-radius: 50%;
+    background-color: rgb(255, 255, 255);
+    color: var(--text-dark-contrast-locked);
+    font-weight: 700;
+
+    &--root {
+      background-color: rgb(255, 0, 0);
+    }
+
+    &--selected {
+      border: var(--text-light-low-emphasis-locked) 0.25rem solid;
+      background-color: var(--text-light-disabled-locked);
+
+      :global([data-theme="dark"]) & {
+        color: var(--text-dark-contrast);
+      }
+    }
+  }
+}
